fix can now import multiplatform

+++ ankama_launcher_emulator/haapi/zaap_version.py
@@ -9,11 +9,13 @@ def get_zaap_version():
             f'asar extract-file "{os.path.join(os.getenv('programfiles', ''), 'Ankama', 'Ankama Launcher','resources', 'app.asar',)}" "{filename}"'
         )
         zaapVersion = "none"
-        with open(filename) as jsonFile:
-            disctJson = json.load(jsonFile)
-            zaapVersion = disctJson.get("version")
+        if os.path.exists(filename):
+            with open(filename) as jsonFile:
+                disctJson = json.load(jsonFile)
+                zaapVersion = disctJson.get("version")
     finally:
-        os.remove(filename)
+        if os.path.exists(filename):
+            os.remove(filename)
 
     if zaapVersion == "none":
         zaapVersion = "3.12.19"
